Crate plutus_datum_derive

This crate creates the attribute macro ToDatum

Derive Macros

ToDatum

Derives ToDatum instance for annotated struct. constructor_datum parameter should be used on a single field structs to decide if the field should be mapped to datum directly or wrapped in constructor datum with one variant and single field. Such distinction is required because exactly same Datum encoding, as in https://github.com/input-output-hk/partner-chains-smart-contracts, is required.